create RegisterStudent, DetailStudent
