+++ src/Components/Account/DetailStudent.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const DetailStudent = () => {
+  return (<div>디테일</div>);
+}
+
+export default DetailStudent;
